Omgggggggg

+++ cell02/ex00/iszero.py
@@ -2,7 +2,7 @@
 
 def main():
     try:
-        num = float(input())
+        num = int(input())
         if num == 0:
             print("This number is equal to zero.")
         else:
